Remove `BuiltinLintDiag::SingleUseLifetime`
